\docType{package}
\name{cdtRggplot2-package}
\alias{cdtRggplot2}
\alias{cdtRggplot2-package}
\title{R courses for Newcastle University's CDT in Cloud Computing for Big Data}
\description{
  Functions and datasets used for Newcastle University's 
  CDT in Cloud Computing for Big Data. This package is 
  used in the "Advanced Graphics" course.
}
\keyword{package}

